+++ app/forms/card_form.py
@@ -4,6 +4,6 @@ from wtforms.validators import DataRequired, Length
 
 class CardForm(FlaskForm):
     name = StringField('Name', validators=[DataRequired(), Length(max=64)])
-    description = StringField('Description', validators=[DataRequired(), Length(max=1000)])
+    description = StringField('Description', validators=[Length(max=1000)])
     list_id = IntegerField('List Id', validators=[DataRequired()])
     submit = SubmitField('Submit')
